reuse search bar widget
