+++ lib/wanderlist/widgets/add_activity.dart
@@ -1,4 +1,5 @@
 import 'package:application/components/activity_summary_item_small.dart';
+import 'package:application/components/searchfield.dart';
 import 'package:application/models/activity.dart';
 import 'package:application/wanderlist/cubit/suggestions_cubit.dart';
 import 'package:application/wanderlist/cubit/suggestions_state.dart'
@@ -34,7 +35,7 @@ class AddActivityOverlay extends StatelessWidget {
             Padding(
               padding: EdgeInsets.only(bottom: 10),
             ),
-            _Search(),
+            SearchField("Search Activities"),
             Padding(
               padding: EdgeInsets.only(bottom: 10),
             ),
@@ -59,31 +60,6 @@ class AddActivityOverlay extends StatelessWidget {
   }
 }
 
-class _Search extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 40,
-      child: TextField(
-        decoration: InputDecoration(
-          contentPadding: EdgeInsets.only(left: 15),
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(25.0),
-            borderSide: BorderSide(
-              width: 0,
-              style: BorderStyle.none,
-            ),
-          ),
-          filled: true,
-          hintStyle: TextStyle(color: Colors.grey[800], height: 1),
-          hintText: "Search activities",
-          fillColor: Colors.grey[300],
-        ),
-        style: TextStyle(height: 1),
-      ),
-    );
-  }
-}
 
 class _Suggestions extends StatelessWidget {
   @override
@@ -181,7 +157,7 @@ class _SuggestionsItem extends StatelessWidget {
           );
         }
 
-        return Text("Impossible state");
+        throw Exception("Impossible state");
       },
       listener: (context, state) {},
     );
